requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
